+++ components/Layout/CommonFinancialInfo.tsx
