+++ components/TestHeader/index.tsx
@@ -1,0 +1,24 @@
+export const TestHeader = () => (
+  <div className="bg-[#063056]">
+    <div className="container mx-auto">
+      <div className="flex items-center">
+        <div className="block lg:flex py-4 layout-container">
+          <div className="flex justify-between lg:block lg:w-max ">
+            <span
+              className="text-white border block w-max px-4 py-1 my-auto leading-6"
+              role="alert"
+            >
+              TEST SITE
+            </span>
+          </div>
+          <div className="lg:ml-4 xl:ml-8 xxl:ml-12">
+            <p className="font-body mt-5 lg:mt-auto text-white lg:ml-4 pt-1 my-auto lg:mb-0 lg:pb-1">
+              You cannot change any information through this test site. This is
+              a POC project.
+            </p>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+)
